OpenWebAuth moved to a separate class / Improved authentication handling
