+++ src/Content/Nav.php
@@ -30,12 +30,12 @@ use Friendica\Core\Renderer;
 use Friendica\Core\Session\Capability\IHandleUserSessions;
 use Friendica\Database\Database;
 use Friendica\Model\Contact;
-use Friendica\Model\Profile;
 use Friendica\Model\User;
 use Friendica\Module\Conversation\Community;
 use Friendica\Module\Home;
 use Friendica\Module\Security\Login;
 use Friendica\Network\HTTPException;
+use Friendica\Security\OpenWebAuth;
 
 class Nav
 {
@@ -281,7 +281,7 @@ class Nav
 
 		$gdirpath = 'directory';
 		if ($this->config->get('system', 'singleuser') && $this->config->get('system', 'directory')) {
-			$gdirpath = Profile::zrl($this->config->get('system', 'directory'), true);
+			$gdirpath = OpenWebAuth::getZrlUrl($this->config->get('system', 'directory'), true);
 		}
 
 		if (Feature::isEnabled($this->session->getLocalUserId(), Feature::COMMUNITY) && (($this->session->getLocalUserId() || $this->config->get('system', 'community_page_style') != Community::DISABLED_VISITOR) &&
